six/requirements.txt
djangorestframework
django
djangorestframework_simplejwt